+++ entrypoint.sh
@@ -1,16 +1,17 @@
 #!/bin/sh
 set -e
 
-if [[ $1 != "${1%bash}" || $1 != "${1%sh}" ]]; then
-  exec "$@"
+if [ "$#" -eq 0 ] || [ "${1#-}" != "$1" ]; then
+  echo dx-scanner run --ci -r $@
+  set -- dx-scanner run --ci -r "$@"
 fi
 
 echo "Run DX Scanner"
 
 if [[ $1 != "${1#dx-scanner}" || $1 != "${1#dxs}" || $1 != "${1#dxscanner}" ]]; then
+  echo  $@
+  exec "$@"
+fi
+
   echo $@
   exec "$@"
-else
-  echo dx-scanner run --ci -r $@
-  exec dx-scanner run --ci -r "$@"
-fi